Sort and search images.
Place images to be sorted(based on size) in images folder. The program will sort the images and place them accordingly in the sorted folder. For searching, place the image(to be searched) in the 'Finder' folder and the program will successfully give you the result. The program also computes the total time take to execute.